Separated distance factor from religion factor in TOR aggression algorithm. Refactored aggression into a diplomacy subsystem
