+++ ModuleData/tor_custom_xmls/tor_config.xml
@@ -11,7 +11,7 @@
     NumberOfInitialHideoutsAtEachBanditFaction="40"
     NumberOfMaximumHideoutsAtEachBanditFaction="80"
     MaximumNumberOfCareerPerkPoints="30"
-    DeclareWarScoreMultiplierTor="8"
+    DeclareWarScoreMultiplierTor="4"
     DeclareWarScoreMultiplierNative="2"
-    DeclarePeaceMultiplier="3"
+    DeclarePeaceMultiplier="2"
 />
